refactor: Access Token을 이용한 내 예약 확인하기 및 전체 스케줄 확인하기 구현
