+++ backend/src/main/java/com/ice/studyroom/domain/reservation/infrastructure/persistence/ReservationRepository.java
@@ -1,4 +1,4 @@
-package com.ice.studyroom.domain.reservation.dao;
+package com.ice.studyroom.domain.reservation.infrastructure.persistence;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -9,4 +9,6 @@ import com.ice.studyroom.domain.reservation.domain.entity.Reservation;
 
 public interface ReservationRepository extends JpaRepository<Reservation, Long> {
 	List<Reservation> findByScheduleDate(LocalDate date);
+
+	List<Reservation> findByUserEmail(String email);
 }
