+++ backend/src/main/java/com/ice/studyroom/domain/reservation/presentation/ReservationController.java
@@ -1,4 +1,4 @@
-package com.ice.studyroom.domain.reservation.api;
+package com.ice.studyroom.domain.reservation.presentation;
 
 import java.util.List;
 
@@ -10,15 +10,16 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.ice.studyroom.domain.reservation.application.ReservationService;
 import com.ice.studyroom.domain.reservation.domain.entity.Reservation;
 import com.ice.studyroom.domain.reservation.domain.entity.Schedule;
-import com.ice.studyroom.domain.reservation.dto.request.CreateReservationRequest;
-import com.ice.studyroom.domain.reservation.dto.request.DeleteReservationRequest;
-import com.ice.studyroom.domain.reservation.dto.response.ReservationResponse;
+import com.ice.studyroom.domain.reservation.presentation.dto.request.CreateReservationRequest;
+import com.ice.studyroom.domain.reservation.presentation.dto.request.DeleteReservationRequest;
+import com.ice.studyroom.domain.reservation.presentation.dto.response.ReservationResponse;
 import com.ice.studyroom.global.dto.response.ResponseDto;
 import com.ice.studyroom.global.exception.BusinessException;
 import com.ice.studyroom.global.type.StatusCode;
@@ -37,33 +38,43 @@ public class ReservationController {
 
 	private final ReservationService reservationService;
 
-	@GetMapping("/schedules")
-	public ResponseEntity<List<Schedule>> getSchedules() {
-		return ResponseEntity
-			.status(HttpStatus.OK)
-			.body(reservationService.getSchedules());
-	}
-
+	/**
+	 *
+	 * @param authorizationHeader
+	 * @return List 형태의 내 예약 정보들
+	 * exception handler 전역 처리로 수정 예정
+	 */
 	@ExceptionHandler(BusinessException.class)
-	@GetMapping("/reservations")
-	public ResponseEntity<ResponseDto<List<Reservation>>> getReservations() {
+	@GetMapping("/reservations/my")
+	public ResponseEntity<ResponseDto<List<Reservation>>> getMyReservation(
+		@RequestHeader("Authorization") String authorizationHeader
+	) {
 		return ResponseEntity
 			.status(StatusCode.OK.getStatus())
-			.body(ResponseDto.of(reservationService.getReservations()));
+			.body(ResponseDto.of(reservationService.getMyReservation(authorizationHeader)));
+	}
+
+	@GetMapping("/schedules")
+	public ResponseEntity<ResponseDto<List<Schedule>>> getSchedule() {
+		return ResponseEntity
+			.status(HttpStatus.OK)
+			.body(ResponseDto.of(reservationService.getSchedule()));
 	}
 
 	@Operation(summary = "스터디룸 예약", description = "스터디룸을 예약합니다.")
 	@ApiResponse(responseCode = "201", description = "스터디룸 예약 성공")
 	@PostMapping("/reservations")
 	public ResponseEntity<ResponseDto<ReservationResponse>> createReservation(
+		@RequestHeader("Authorization") String authorizationHeader,
 		@Valid @RequestBody CreateReservationRequest request
 	) {
 		return ResponseEntity
 			.status(StatusCode.OK.getStatus())
-			.body(ResponseDto.of(reservationService.createReservation(request)));
+			.body(ResponseDto.of(reservationService.createReservation(authorizationHeader, request)));
 	}
 
-	@Operation(summary = "예약 삭제", description = "예약을 삭제합니다.")
+	// 예약 취소 시 본인 인증이 필요하다.
+	@Operation(summary = "예약 취소", description = "예약을 취소합니다.")
 	@DeleteMapping("/reservations/{id}")
 	public ResponseEntity<ResponseDto<Void>> deleteReservation(@PathVariable Long id) {
 		DeleteReservationRequest request = DeleteReservationRequest.builder()
